pyproject.toml
[project]
name = "raguardiannews"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "bertopic>=0.17.4",
    "duckdb>=1.4.4",
    "faiss-cpu>=1.13.2",
    "google>=3.0.0",
    "google-genai>=1.73.1",
    "hdbscan>=0.8.42",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.4.0",
    "spacy>=3.8.11",
    "streamlit>=1.56.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
    "umap-learn>=0.5.12",
]
